+++ internal/testutils/test_utils.go
@@ -86,7 +86,7 @@ func GetSamplePrivKey() *rsa.PrivateKey {
 }
 
 func SetupMockDataLoader(subscribed []string, result []authservice.ServiceReg, err error) *mocks.AuthDataLoader {
-	mockLoader := mocks.NewAuthDataLoader("https://auth.rokwire.com", subscribed)
+	mockLoader := mocks.NewAuthDataLoader("https://auth.rokwire.com", "/service-regs", subscribed)
 	mockLoader.On("LoadServices").Return(result, err)
 	return mockLoader
 }
@@ -100,7 +100,7 @@ func SetupExampleMockLoader() *mocks.ServiceRegLoader {
 	authServiceReg := authservice.ServiceReg{ServiceID: "auth", Host: "https://auth.rokwire.com", PubKey: GetSamplePubKey()}
 	serviceRegsValid := []authservice.ServiceReg{authServiceReg, testServiceReg}
 
-	mockLoader := mocks.NewServiceRegLoader(nil)
+	mockLoader := mocks.NewServiceRegLoader("/service-regs", nil)
 	mockLoader.On("LoadServices").Return(serviceRegsValid, nil)
 
 	return mockLoader